+++ src/unpoly-migrate/form.js
@@ -9,9 +9,9 @@ up.migrate.renamedProperty(up.form.config, 'observeDelay', 'watchInputDelay')
 
 up.migrate.migratedFormGroupSelectors = function() {
   return up.form.config.groupSelectors.map((originalSelector) => {
-    let migratedSelector = originalSelector.replace(/:has\((&|:origin)\)$/, '')
+    let migratedSelector = originalSelector.replace(/:has\(:origin\)$/, '')
     if (originalSelector !== migratedSelector) {
-      up.migrate.warn('Selectors in up.form.config.groupSelectors must not contain ":has(&)". The suffix is added automatically where required. Found in "%s".', originalSelector)
+      up.migrate.warn('Selectors in up.form.config.groupSelectors must not contain ":has(:origin)". The suffix is added automatically where required. Found in "%s".', originalSelector)
     }
     return migratedSelector
   })